Fix prose plugin not picking up theme colors
Because tailwind switched to using CSS variables but we were still
loading the colors from the object, the prose plugin wasn't picking up
the theme colors, meaning all the links were black.

However, we couldn't just switch to using a css file for the theme
values because we need them from the javascript to configure the ol
library. And because we want the values at SSR time, we couldn't easily
just read the CSS variables.

The compromise is to add a script to generate the theme.css file that
can be used by tailwind from the colors defined in the theme object.